refactor: 사다리 검증 책임을 분리
