+++ src/test/java/domain/HeightTest.java
@@ -1,0 +1,16 @@
+package domain;
+
+import static org.assertj.core.api.Assertions.assertThatThrownBy;
+
+import org.junit.jupiter.api.DisplayName;
+import org.junit.jupiter.api.Test;
+
+class HeightTest {
+
+    @DisplayName("사다리의 높이는 1 이상만 허용한다.")
+    @Test
+    void checkLadderHeight() {
+        assertThatThrownBy(() -> new Height(0))
+                .isInstanceOf(IllegalArgumentException.class);
+    }
+}
