+++ src/test/java/domain/LadderTest.java
@@ -2,7 +2,6 @@ package domain;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatCode;
-import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 import java.util.List;
 import org.junit.jupiter.api.DisplayName;
@@ -15,13 +14,6 @@ class LadderTest {
     void create() {
         assertThatCode(() -> Ladder.createRandomLadder(3, 2))
                 .doesNotThrowAnyException();
-    }
-
-    @DisplayName("사다리의 높이는 1 이상만 허용한다.")
-    @Test
-    void checkLadderHeight() {
-        assertThatThrownBy(() -> Ladder.createRandomLadder(0, 2))
-                .isInstanceOf(IllegalArgumentException.class);
     }
 
     @DisplayName("사다리는 높이 만큼의 라인을 갖는다.")
